Modularise triehash encodiung further
