+++ packages/util-triehash/src/encode/hexPrefix.js
